Encapsulated fncAskForOffset to reduce code redundancy
I was concerned that there was a bug in the repeated code, so I took the
version that worked, and made a function out of it. This is maybe not
the most efficient implementation since it re-parses the whole calc file
if you just want to update the offset, but it's not a big delay on my
computer. Could be fixed in future updates.