Plot format changes
Added a text box with the passing rate onto the gamma plots.
Included the calc dose file name in the title of each plot.
Included the calc dose file name in the name of the pdf file.
Changed the positioning and size of the plots as they are displayed at
runtime.